requirements.txt
opencv-python==4.0.0.21
azure-cognitiveservices-vision-customvision==1.0.0
